+++ pytorch_lightning/trainer/connectors/checkpoint_connector.py
@@ -258,16 +258,23 @@ class CheckpointConnector:
 
     def restore_optimizers_and_schedulers(self) -> None:
         """Restores the optimizers and learning rate scheduler states from the pre-loaded checkpoint."""
-        if not self._loaded_checkpoint or not self.trainer.strategy.lightning_restore_optimizer_and_schedulers:
+        if not self._loaded_checkpoint:
             return
 
-        # validation
-        if "optimizer_states" not in self._loaded_checkpoint or "lr_schedulers" not in self._loaded_checkpoint:
+        if self.trainer.strategy.lightning_restore_optimizer:
+            # validation
+            if "optimizer_states" not in self._loaded_checkpoint:
+                raise KeyError(
+                    "Trying to restore optimizer state but checkpoint contains only the model."
+                    " This is probably due to `ModelCheckpoint.save_weights_only` being set to `True`."
+                )
+            self.restore_optimizers()
+
+        if "lr_schedulers" not in self._loaded_checkpoint:
             raise KeyError(
-                "Trying to restore training state but checkpoint contains only the model."
+                "Trying to restore learning rate scheduler state but checkpoint contains only the model."
                 " This is probably due to `ModelCheckpoint.save_weights_only` being set to `True`."
             )
-        self.restore_optimizers()
         self.restore_lr_schedulers()
 
     def restore_optimizers(self) -> None: